python_scripts/ops/requirements.txt
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
Django==2.2.6
django-extensions==2.2.5
ice==0.0.2
idna==2.8
IPy==1.0
jieba==0.42.1
Jinja2==2.10.3
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.17.3
pandas==0.25.2
paramiko==2.6.0
Pillow==7.1.2
pipenv==2018.11.26
pyasn1==0.4.8
pycparser==2.19
pymongo==3.9.0
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
redis==3.5.3
requests==2.22.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
 tencentcloud-sdk-python==3.0.259
urllib3==1.25.6
virtualenv==16.7.6
virtualenv-clone==0.5.3
zeroc-ice==3.7.4